refactor(frontend): replace hardcoded slug strings with Slugs const
Updates all frontend pages and helpers to use Slugs.Collections.*
and Slugs.Globals.* instead of raw strings in find/findGlobal calls.
