+++ src/app/(frontend)/_components/SponsorsServerSection.tsx
@@ -1,10 +1,10 @@
 import { SponsorsSection } from "@/components/Composite"
-import { payload } from "@/lib/payload"
+import { payload, Slugs } from "@/lib/payload"
 import type { Sponsor } from "@/payload/payload-types"
 
 export const SponsorsServerSection = async () => {
   const { docs: sponsors }: { docs: Sponsor[] } = await payload.find({
-    collection: "sponsor",
+    collection: Slugs.Collections.sponsor,
   })
 
   return <SponsorsSection sponsors={sponsors} />
