+++ src/app/(frontend)/sponsors/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next"
 import { Container, Heading, LazyImage } from "@/components/Primitive"
-import { payload } from "@/lib/payload"
+import { payload, Slugs } from "@/lib/payload"
 import { cn } from "@/lib/utils"
 import type { Sponsor } from "@/payload/payload-types"
 import { SPONSOR_TIER_ORDER, SponsorTier } from "@/types/enums"
@@ -13,7 +13,7 @@ export const metadata: Metadata = {
 
 export default async function SponsorsPage() {
   const { docs: sponsors }: { docs: Sponsor[] } = await payload.find({
-    collection: "sponsor",
+    collection: Slugs.Collections.sponsor,
     depth: 2,
   })
