+++ src/lib/helpers.ts
@@ -1,9 +1,10 @@
 import type { SocialLink } from "@/components/Generic"
 import { payload } from "./payload"
+import { Slugs } from "./slugs"
 
 export async function getSocialLinks(): Promise<SocialLink[]> {
   const { discordHref, instagramHref, tiktokHref, linkedinHref } = await payload.findGlobal({
-    slug: "social-links",
+    slug: Slugs.Globals.socialLinks,
   })
   return [
     { icon: "discord", label: "Discord", href: discordHref ?? "" },
